feat(design-system): establish portable open-source foundations
