+++ .github/workflows/design-system.yml
@@ -1,0 +1,19 @@
+name: Design system
+on:
+  pull_request:
+    paths: ['desktop-next/**', 'packages/**', 'pnpm-lock.yaml', 'pnpm-workspace.yaml', '.github/workflows/design-system.yml']
+  push:
+    branches: [design-system]
+    paths: ['desktop-next/**', 'packages/**', 'pnpm-lock.yaml', 'pnpm-workspace.yaml', '.github/workflows/design-system.yml']
+permissions:
+  contents: read
+jobs:
+  design-system:
+    runs-on: ubuntu-latest
+    timeout-minutes: 15
+    steps:
+      - uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10
+      - uses: cashapp/activate-hermit@cea9af7913204a965fd488637a8d1811bba2e616
+      - run: pnpm install --filter buzz-desktop-next... --frozen-lockfile
+      - run: pnpm --filter buzz-desktop-next check
+      - run: pnpm --filter buzz-desktop-next build
