+++ desktop-next/src/features/design-system/ui/SpacingPage.tsx
@@ -1,24 +1,20 @@
-import { PageHeader, Section, Stub } from "./primitives";
-
+import { FOUNDATION_GROUPS } from "@buzz/design-tokens/foundations";
+import { PageHeader, Section } from "./primitives";
 export function SpacingPage() {
+  const group = FOUNDATION_GROUPS[0];
   return (
     <>
-      <PageHeader
-        title="Spacing"
-        status="not started"
-        intro="Tailwind's default spacing scale is currently in play, unexamined. Whether that is the right rhythm for this design is an open decision."
-      />
-
-      <Section title="What this page will hold">
-        <Stub
-          what="A spacing scale with the same two-layer treatment as colour: raw steps below, and named roles for the spacings that carry meaning — the gutter between rows, the inset of a panel, the gap inside a control."
-          decide={[
-            "Whether the stock scale stays or the design needs its own rhythm",
-            "Which spacings are named roles rather than raw steps",
-            "Whether density varies per surface, and if so how that is expressed",
-            "The relationship between spacing and the type ramp",
-          ]}
-        />
+      <PageHeader title={group.name} intro={group.description} />
+      <Section title="Roles">
+        <dl className="grid gap-6">
+          {group.roles.map(([name, value, use]) => (
+            <div key={name} className="grid gap-2 sm:grid-cols-3">
+              <dt className="text-code font-mono">{name}</dt>
+              <dd className="text-body text-secondary">{value}</dd>
+              <dd className="text-caption text-secondary">{use}</dd>
+            </div>
+          ))}
+        </dl>
       </Section>
     </>
   );
