+++ desktop-next/src/main.tsx
@@ -1,3 +1,5 @@
+import "@fontsource-variable/inter/index.css";
+import "@fontsource/jetbrains-mono/index.css";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { RouterProvider, createRouter } from "@tanstack/react-router";
