+++ packages/design-tokens/build.mjs
@@ -1,0 +1,14 @@
+import { readFileSync, writeFileSync } from "node:fs";
+import { FOUNDATION_GROUPS } from "./src/foundations.ts";
+const declarations = FOUNDATION_GROUPS.flatMap((g) =>
+  g.roles.map(([name, value]) => "  --" + name + ": " + value + ";"),
+).join("\n");
+const css =
+  "/* Generated by packages/design-tokens/build.mjs. */\n@theme {\n" +
+  declarations +
+  "\n}\n@media (prefers-reduced-motion: reduce) {\n  :root { --duration-state: 0ms; --duration-enter: 0ms; --duration-exit: 0ms; --press-scale: 1; }\n}\n";
+const path = new URL("./src/foundations.css", import.meta.url);
+if (process.argv.includes("--check")) {
+  if (readFileSync(path, "utf8") !== css)
+    throw new Error("Foundation CSS is stale: run pnpm tokens:build");
+} else writeFileSync(path, css);
