+++ packages/design-tokens/src/foundations.css
@@ -1,0 +1,23 @@
+/* Generated by packages/design-tokens/build.mjs. */
+@theme {
+  --space-control-gap: 0.5rem;
+  --space-control-inset: 0.75rem;
+  --space-section-gap: 1.5rem;
+  --space-panel-inset: 1.5rem;
+  --space-page-inset: 2rem;
+  --radius-control: 0.5rem;
+  --radius-notice: 0.75rem;
+  --radius-menu: 1rem;
+  --radius-container: 1.5rem;
+  --radius-overlay: 2rem;
+  --radius-sheet: 2.5rem;
+  --radius-pill: 9999px;
+  --duration-state: 120ms;
+  --duration-enter: 180ms;
+  --duration-exit: 120ms;
+  --ease-standard: cubic-bezier(0.2, 0, 0, 1);
+  --press-scale: 0.97;
+}
+@media (prefers-reduced-motion: reduce) {
+  :root { --duration-state: 0ms; --duration-enter: 0ms; --duration-exit: 0ms; --press-scale: 1; }
+}
