+++ pnpm-workspace.yaml
@@ -1,4 +1,5 @@
 packages:
+  - "packages/*"
   - "desktop"
   - "desktop-next"
   - "web"
